+++ 2DGame/2DGame/entities/Wall.java
@@ -4,7 +4,7 @@ import java.awt.Color;
 
 public class Wall extends Entity{
 
-	public Wall(int x, int y,int w, int h) {
+	public Wall(double x, double y,int w, int h) {
 		super(x, y, Color.blue);
 		this.width = w;
 		this.height = h;
